requirements.txt
pyyaml
dataclasses
typing